rename & divide file
